+++ anubis-management-api/anubis/wac.py
@@ -1,6 +1,6 @@
 from rdflib import Graph, URIRef, BNode, Literal
 from rdflib import Namespace
-from rdflib.namespace import FOAF, RDF
+from rdflib.namespace import FOAF, RDF, NamespaceManager
 from anubis.policies.models import Policy
 from sqlalchemy.orm import Session
 import yaml
@@ -9,30 +9,44 @@ import os
 acl = Namespace("http://www.w3.org/ns/auth/acl#")
 oc_acl = Namespace("http://voc.orchestracities.io/oc-acl#")
 
-# follow keycloak scheme
-# user urls should be /{realm}/users/{id}
-# group urls should be /{realm}/groups/{id}
-
 
 def parse_rdf_graph(data):
     g = Graph()
+    g.bind("acl", acl)
+    g.bind("oc_acl", oc_acl)
     g.parse(data=data)
+    nm = NamespaceManager(g)
     policies = {}
     for subj, pred, obj in g:
         policy = str(subj).split("/")[-1]
         if not policies.get(policy):
             policies[policy] = {}
         if pred == acl.agentClass:
-            policies[policy]["agentClass"] = str(obj)
+            policies[policy]["agentClass"] = redux(str(nm.normalizeUri(obj)))
         if pred == acl.mode:
-            policies[policy]["mode"] = str(obj)
+            policies[policy]["mode"] = redux(str(nm.normalizeUri(obj)))
+        # TODO deal with this better.
         if pred == acl.accessTo:
-            policies[policy]["accessTo"] = str(obj).split("://")[-1]
+            policies[policy]["accessTo"] = str(obj).split("/")[-1]
         if pred == acl.default:
             policies[policy]["accessTo"] = "default"
         if pred == acl.accessToClass:
-            policies[policy]["accessToClass"] = str(obj).split("/")[-1]
+            policies[policy]["accessToClass"] = str(
+                nm.normalizeUri(obj)).split(":")[-1]
     return policies
+
+
+def redux(uri: str):
+    if "http://www.w3.org/ns/auth/acl#" in uri:
+        return "acl:" + uri[len("http://www.w3.org/ns/auth/acl#") + 1:-1]
+    if "http://voc.orchestracities.io/oc-acl#" in uri:
+        return "oc-acl:" + \
+            uri[len("http://voc.orchestracities.io/oc-acl#") + 1:-1]
+    if "http://xmlns.com/foaf/0.1/" in uri:
+        return "foaf:" + uri[len("http://xmlns.com/foaf/0.1/") + 1:-1]
+    if "<" in uri[0] and ">" in uri[-1]:
+        return uri[1:-1]
+    return uri
 
 
 def serialize(