+++ scripts/test_context_broker.sh
@@ -2,8 +2,14 @@
 
 echo "Obtaining token from Keycloak..."
 
-export token=`curl -s -d "client_id=client1&grant_type=password&username=admin&password=admin" -X POST --header "Host: keycloak:8080" 'http://localhost:8080/auth/realms/master/protocol/openid-connect/token' | \
-jq -j '.access_token'`
+export json=$(curl -s -d "client_id=ngsi&client_secret=changeme&grant_type=password&username=admin@mail.com&password=admin" -X POST --header "Host: keycloak:8080" 'http://localhost:8080/realms/default/protocol/openid-connect/token')
+export token=$( jq -r ".access_token" <<<"$json" )
+
+echo ""
+echo "decoded token:"
+echo ""
+
+jq -R 'split(".") | .[1] | @base64d | fromjson' <<< $( jq -r ".access_token" <<<"$json" )
 
 echo "Can I read all entities?"
 echo "==============================================================="